I was just cleaning some things up and removing all the redundant parameters and confusing names then I found a really bad bug.
When we switched to multithreading we were still using the distance
attribute of the training data attributes, resetting them each time
through the predict function. Thus all threads were racing to set and
use the distance attribute of the recipes. 

I fixed this by using a local array of doubles to store distances and
changing the nearest neighbors array to store indices into the
trainingData rather than the recipe object itself.

I havent let it run to completion yet but it appears we are now getting
88% to 89% with just the jaccard distance and vote weighting inversely
proportional to distance. 

I can't find anything wrong with it, the code for checking duplicates
hasn't changes so it shouldn't be overfitting again...